allow floats

+++ utils/add_media.py
@@ -10,7 +10,7 @@ config.read('config.ini')
 
 
 @click.command()
-@click.argument('ml', type=int)
+@click.argument('ml', type=float)
 def add_media(ml):
     GPIO.setmode(GPIO.BCM)
